+++ R/GitStats.R
@@ -13,17 +13,21 @@ GitStats <- R6::R6Class("GitStats",
     #' @param repos An optional character vector of repositories full names
     #'   (organization and repository name, e.g. "r-world-devs/GitStats"). If
     #'   you pass it, `orgs` parameter should stay `NULL`.
+    #' @param verbose A logical, `TRUE` by default. If `FALSE` messages and printing
+    #'   output is switched off.
     #' @return Nothing, puts connection information into `$hosts` slot.
     set_github_host = function(host,
                                token = NULL,
                                orgs = NULL,
-                               repos = NULL) {
+                               repos = NULL,
+                               verbose = TRUE) {
       new_host <- NULL
       new_host <- GitHostGitHub$new(
         orgs = orgs,
         repos = repos,
         token = token,
-        host = host
+        host = host,
+        verbose = verbose
       )
       private$add_new_host(new_host)
     },
@@ -38,17 +42,21 @@ GitStats <- R6::R6Class("GitStats",
     #' @param repos An optional character vector of repositories full names
     #'   (organization and repository name, e.g. "r-world-devs/GitStats"). If
     #'   you pass it, `orgs` parameter should stay `NULL`.
+    #' @param verbose A logical, `TRUE` by default. If `FALSE` messages and printing
+    #'   output is switched off.
     #' @return Nothing, puts connection information into `$hosts` slot.
     set_gitlab_host = function(host,
                                token = NULL,
                                orgs = NULL,
-                               repos = NULL) {
+                               repos = NULL,
+                               verbose = TRUE) {
       new_host <- NULL
       new_host <- GitHostGitLab$new(
         orgs = orgs,
         repos = repos,
         token = token,
-        host = host
+        host = host,
+        verbose = verbose
       )
       private$add_new_host(new_host)
     },