+++ R/gitstats_functions.R
@@ -19,6 +19,8 @@ create_gitstats <- function() {
 #' @param repos An optional character vector of repositories full names
 #'   (organization and repository name, e.g. "r-world-devs/GitStats"). If you
 #'   pass it, `orgs` parameter should stay `NULL`.
+#' @param verbose A logical, `TRUE` by default. If `FALSE` messages and printing
+#'   output is switched off.
 #' @details If you do not define `orgs` and `repos`, `GitStats` will be set to
 #'   scan whole Git platform (such as enterprise version of GitHub or GitLab),
 #'   unless it is a public platform. In case of a public one (like GitHub) you
@@ -37,12 +39,14 @@ set_github_host <- function(gitstats_object,
                             host = NULL,
                             token = NULL,
                             orgs = NULL,
-                            repos = NULL) {
+                            repos = NULL,
+                            verbose = is_verbose(gitstats_object)) {
   gitstats_object$set_github_host(
     host = host,
     token = token,
     orgs = orgs,
-    repos = repos
+    repos = repos,
+    verbose = verbose
   )
 
   return(invisible(gitstats_object))
@@ -70,12 +74,14 @@ set_gitlab_host <- function(gitstats_object,
                             host = NULL,
                             token = NULL,
                             orgs = NULL,
-                            repos = NULL) {
+                            repos = NULL,
+                            verbose = is_verbose(gitstats_object)) {
   gitstats_object$set_gitlab_host(
     host = host,
     token = token,
     orgs = orgs,
-    repos = repos
+    repos = repos,
+    verbose = verbose
   )
 
   return(invisible(gitstats_object))